+++ src/models/modules/session.ts
@@ -45,11 +45,6 @@ export default class Session extends Provider {
 	 */
 	readonly csrfHeader: string = 'X-XSRF-TOKEN';
 
-	/**
-	 * If true, then after clearing the session (in case of answer 401) will be an additional query
-	 */
-	readonly requestAfterClear: boolean = true;
-
 	/** @override */
 	static readonly middlewares: Middlewares = {
 		...Provider.middlewares,
@@ -78,30 +73,36 @@ export default class Session extends Provider {
 	}
 
 	/** @override */
-	protected updateRequest<T = unknown>(url: string, factory: RequestFunctionResponse<T>): RequestResponse<T>;
+	protected updateRequest<T = unknown>(
+		url: string,
+		factory: RequestFunctionResponse<T>
+	): RequestResponse<T>;
+
 	protected updateRequest<T = unknown>(
 		url: string,
 		event: string,
-		factory: RequestFunctionResponse<T>
+		factory: RequestFunctionResponse<T>,
+		canRetry?: boolean
 	): RequestResponse<T>;
 
 	protected updateRequest(
 		url: string,
 		event: string | RequestFunctionResponse,
-		factory?: RequestFunctionResponse
+		factory?: RequestFunctionResponse,
+		canRetry: boolean = true
 	): RequestResponse {
 		const
 			req = super.updateRequest(url, <any>event, <any>factory),
-			session = s.get();
+			sessionPromise = s.get();
 
 		const update = async (res) => {
 			const
-				info = <Response>res.response,
-				refreshHeader = info.getHeader(this.authRefreshHeader);
+				{response} = res,
+				refreshHeader = response.getHeader(this.authRefreshHeader);
 
 			try {
 				if (refreshHeader != null) {
-					await s.set(refreshHeader, {csrf: info.getHeader(this.csrfHeader)});
+					await s.set(refreshHeader, {csrf: response.getHeader(this.csrfHeader)});
 				}
 
 			} catch {}
@@ -109,27 +110,27 @@ export default class Session extends Provider {
 
 		req
 			.then(update)
-			.catch(stderr);
+			// logging must be already handled in a request factory. Just do nothing
+			.catch(() => {});
 
 		return req.catch(async (err) => {
 			const
-				response = Object.get<Response>(err, 'details.response'),
-				{auth, params} = await session;
+				response = Object.get<Response>(err, 'details.response');
 
 			if (response) {
-				const
-					r = () => this.updateRequest(url, <string>event, <RequestFunctionResponse>factory);
+				if (
+					response.status === 401 &&
+					canRetry &&
+					await s.isExists()
+				) {
+					const
+						{auth, params} = await sessionPromise;
 
-				if (response.status === 401) {
 					if (!await s.match(auth, params)) {
-						return r();
+						return this.updateRequest(url, <string>event, <RequestFunctionResponse>factory, false);
 					}
 
 					await s.clear();
-
-					if (Object.isTruly(auth) && this.requestAfterClear) {
-						return r();
-					}
 				}
 
 				await update({response});